+++ CHANGES.rst
@@ -18,6 +18,11 @@ Changelog
    Commit items for NEXT VERSION
    ..............................
 
+   - fix(conftest): in clean environment pytest_report_header avoid import error
+   - tests: complete static typing
+   - chore: bump versions
+   - ci: bump versions
+
 .. scriv-start-here
 
 .. _changes_0-2-2: